requirements.txt
Flask==2.3.2
Flask-Cors~=3.0.10
Flask-JWT-Extended~=4.4.4
Flask-Migrate~=4.0.4
Flask-SQLAlchemy~=3.0.3
Flask-Limiter==3.3.1
Flask-Bcrypt==1.0.1
mysqlclient~=2.1.1
marshmallow~=3.19.0
python-dotenv~=1.0.0
gunicorn==20.1.0
eventlet==0.33.3
redis~=4.5.4
gevent~=22.10.2